requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
matplotlib>=3.7.0
tensorboard>=2.14.0
tqdm>=4.66.0
einops>=0.7.0
trimesh>=4.0.0
open3d>=0.17.0
face-alignment>=1.4.0
facenet-pytorch>=2.5.3
lpips>=0.1.4
PyMCubes>=0.1.4
kornia>=0.7.0
timm>=0.9.0
albumentations>=1.3.0
wandb>=0.15.0
gdown>=4.7.0
kaggle>=1.5.0
